+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixes
 
 - Correctly install asset folder in Windows installer by @garyttierney
+- Fix mod host crash handler being uninstalled as soon as it was created by @Dasaav-dsv in <https://github.com/garyttierney/me3/pull/91>
 
 ## [v0.4.0] - 2025-06-07